Refactorize lines commission to invoice
delete lines commission table, model and views
link settled commission to customer invoice
create model view and views
