--- Core/Table/lineasliquidacioncomision.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-    Document   : lineasliquidacioncomision.xml
-    Author     : Artex Trading sa <jcuello@artextrading.com>
-    Description: Estructura para la tabla de lineas de liquidacion de comisiones.
--->
-
-
-<table>
-    <column>
-        <name>idfactura</name>
-        <type>integer</type>
-        <null>NO</null>
-    </column>
-    <column>
-        <name>idliquidacion</name>
-        <type>integer</type>
-        <null>NO</null>
-    </column>
-    <constraint>
-        <name>lineasliquidacioncomision_pkey</name>
-        <type>PRIMARY KEY (idfactura)</type>
-    </constraint>
-    <constraint>
-        <name>ca_lineasliquidacioncomision_facturas</name>
-        <type>FOREIGN KEY (idfactura) REFERENCES facturascli (idfactura) ON DELETE RESTRICT ON UPDATE CASCADE</type>
-    </constraint>
-    <constraint>
-        <name>ca_lineasliquidacioncomision_liquidacion</name>
-        <type>FOREIGN KEY (idliquidacion) REFERENCES settledcommissions (idliquidacion) ON DELETE CASCADE ON UPDATE CASCADE</type>
-    </constraint>
-</table>
