+++ Core/XMLView/ListLiquidacionComisionFactura.xml
@@ -1,0 +1,101 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+ * This file is part of FacturaScripts
+ * Copyright (C) 2017-2019  Carlos Garcia Gomez  <carlos@facturascripts.com>
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ *
+ *
+ * Initial description for List Settled Invoice
+ *
+ * @author Artex Trading sa <jcuello@artextrading.com>
+-->
+
+
+<view>
+    <columns>
+        <column name="company" display="none" order="100">
+            <widget type="text" fieldname="empresa" />
+        </column>
+        <column name="exercise" display="none" order="110">
+            <widget type="text" fieldname="ejercicio" />
+        </column>
+        <column name="code" order="120">
+            <widget type="text" fieldname="codigo" />
+        </column>
+        <column name="number" display="none" order="130">
+            <widget type="text" fieldname="numero" />
+        </column>
+        <column name="number2" display="none" order="130">
+            <widget type="text" fieldname="numero2" />
+        </column>
+        <column name="customer" order="140">
+            <widget type="text" fieldname="cliente" />
+        </column>
+        <column name="date" display="center" order="150">
+            <widget type="date" fieldname="fecha" />
+        </column>
+        <column name="amount" display="right" order="160">
+            <widget type="money" fieldname="neto">
+                <option color="danger">lt:0</option>
+            </widget>
+        </column>
+        <column name="total" display="none" order="170">
+            <widget type="money" fieldname="total">
+                <option color="danger">lt:0</option>
+            </widget>
+        </column>
+        <column name="commission" display="right" order="180">
+            <widget type="number" decimal="2" fieldname="porcomision" />
+        </column>
+    </columns>
+
+    <modals>
+        <group name="insertinvoices" title="add-invoices" icon="fas fa-file-invoice">
+            <column name="idsettled" display="none" order="0">
+                <widget type="text" fieldname="idliquidacion" readonly="true" />
+            </column>
+            <column name="idexercise" display="none" order="0">
+                <widget type="text" fieldname="codejercicio" readonly="true" />
+            </column>
+            <column name="idagent" display="none" order="0">
+                <widget type="text" fieldname="codagente" readonly="true" />
+            </column>
+            <column name="from-date" numcolumns="6" order="0">
+                <widget type="date" fieldname="datefrom" />
+            </column>
+            <column name="to-date" numcolumns="6" order="0">
+                <widget type="date" fieldname="dateto" />
+            </column>
+            <column name="status" numcolumns="6" order="0">
+                <widget type="select" fieldname="status" required="true" translate="true">
+                    <values title="only-charged">CHARGED</values>
+                    <values title="all">ALL</values>
+                </widget>
+            </column>
+            <column name="domiciled" numcolumns="6" order="0">
+                <widget type="select" fieldname="domiciled" required="true" translate="true">
+                    <values title="all">ALL</values>
+                    <values title="domiciled">DOMICILED</values>
+                    <values title="without-domicile">WITHOUT</values>
+                </widget>
+            </column>
+            <column name="customer" numcolumns="12" order="110">
+                <widget type="autocomplete" fieldname="codcliente">
+                    <values source="clientes" fieldcode="codcliente" fieldtitle="nombre"></values>
+                </widget>
+            </column>
+        </group>
+    </modals>
+</view>
